requirements.txt
python-dotenv
openai
fastapi
pyngrok
nest-asyncio
twilio
mysql-connector-python
mcp
